benchmarks dont lie, changed element wise operations to plain loops
Former-commit-id: a9a812380d5f74c5a29502f82c10e468f9990513 [formerly a190e5c370c00d39884d7876db39894858259063]
Former-commit-id: aa105f38ef6ae3ad36f63784e43909d4f2959a57

+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/BaseTwoArrayElementWiseOp.java
@@ -34,12 +34,12 @@ public abstract  class BaseTwoArrayElementWiseOp extends BaseElementWiseOp imple
     public void applyTransformToDestination(INDArray from,INDArray destination,INDArray other,int i) {
         if(scalarValue != null) {
             if(destination instanceof IComplexNDArray) {
-                IComplexNumber number = (IComplexNumber) apply(destination, (float) scalarValue, i);
+                IComplexNumber number = (IComplexNumber) apply(destination,  scalarValue, i);
                 IComplexNDArray c2 = (IComplexNDArray) destination;
                 c2.putScalar(i,number);
             }
             else {
-                float f = (float)  apply(from, (float) scalarValue, i);
+                float f = (float)  apply(from,  scalarValue, i);
                 destination.putScalar(i,f);
             }
